make GameClient compile

+++ Generals/Code/GameEngine/Include/GameClient/WindowVideoManager.h
@@ -152,7 +152,7 @@ private:
 	size_t operator()(ConstGameWindowPtr p) const
 	{
 		std::hash<UnsignedInt> hasher;
-		return hasher((UnsignedInt)p);
+		return hasher((UnsignedInt)(uintptr_t)p);
 	}
 	};
